Make the process of parsing the default packages more reliable

+++ run.sh
@@ -2,14 +2,14 @@
 
 set -e
 
+source include/common.sh
+source config/"$(get_platform_name)".sh
+
 build() {
   for package in "$@"
   do
-      major=${package%/*}
-      minor=${package#*/}
-      if [ ${minor} == ${major} ]; then
-        minor=""
-      fi
+      major="$(get_dep_major "$package")"
+      minor="$(get_dep_minor "$package")"
       if [ -f ./build-${major}.sh ]; then
         echo "./build-${major}.sh ${minor}"
         ./build-${major}.sh ${minor}
@@ -20,19 +20,14 @@ build() {
 clean() {
   for package in "$@"
   do
-      major=${package%/*}
-      minor=${package#*/}
-      if [ ${minor} == ${major} ]; then
-        minor=""
-      fi
+      major="$(get_dep_major "$package")"
+      minor="$(get_dep_minor "$package")"
       if [ -f ./build-${major}.sh ]; then
         : ${GIS_INSTALL_ROOT:?} ${major:?}
-        rm -r ${GIS_INSTALL_ROOT}/${major}/${minor}
-        rm -r ${GIS_INSTALL_ROOT}/modulefiles/${major}/${minor}
+        rm -r ${GIS_INSTALL_ROOT:?}/${major:?}/${minor}
+        rm -r ${GIS_INSTALL_ROOT:?}/modulefiles/${major:?}/${minor}
       fi
   done
 }
 
-source include/common.sh
-source config/"$(get_platform_name)".sh
 "$@"
